src/main/java/requests/FindBook.java
package requests;

import io.restassured.response.Response;

import static io.restassured.RestAssured.given;

public class FindBook {
    public Response retrieve(String queryName, String queryString) {
        Response response = given()
                .header("Content-type", "application/json")
                .param(queryName, queryString)
                .when()
                .get("/Library/GetBook.php")
                .then()
                .extract().response();
        return response;
    }
}
